refactor to ibc component module
